+++ contracts/RariFundController.sol
@@ -154,6 +154,8 @@ contract RariFundController is Ownable {
      * @param newContract The address of the new RariFundController contract.
      */
     function upgradeFundController(address payable newContract) external onlyOwner {
+        // Verify fund is disabled + verify new fund controller contract
+        require(_fundDisabled, "This fund controller contract must be disabled before it can be upgraded.");
         require(RariFundController(newContract).IS_RARI_FUND_CONTROLLER(), "New contract does not have IS_RARI_FUND_CONTROLLER set to true.");
 
         for (uint256 i = 0; i < _supportedCurrencies.length; i++) {
@@ -426,7 +428,7 @@ contract RariFundController is Ownable {
      * @param all Boolean indicating if all funds are being withdrawn.
      */
     function withdrawFromPoolOptimized(LiquidityPool pool, string calldata currencyCode, uint256 amount, bool all) external fundEnabled onlyManager {
-        all && (pool == LiquidityPool.dYdX || pool == LiquidityPool.mStable) ? _withdrawAllFromPool(pool, currencyCode) : _withdrawFromPool(pool, currencyCode, amount);
+        all ? _withdrawAllFromPool(pool, currencyCode) : _withdrawFromPool(pool, currencyCode, amount);
         if (all) _poolsWithFunds[currencyCode][uint8(pool)] = false;
     }
 
@@ -554,7 +556,7 @@ contract RariFundController is Ownable {
 
         if (inputErc20Contract != address(0)) {
             inputFilledAmountUsd = toUsd(inputCurrencyCode, filledAmounts[0], pricesInUsd);
-            outputFilledAmountUsd = toUsd(inputCurrencyCode, filledAmounts[1], pricesInUsd);
+            outputFilledAmountUsd = toUsd(outputCurrencyCode, filledAmounts[1], pricesInUsd);
             handleExchangeLoss(inputFilledAmountUsd, outputFilledAmountUsd, rawFundBalanceBeforeExchange);
         }
 
@@ -656,5 +658,15 @@ contract RariFundController is Ownable {
 
         // Emit event
         emit CurrencyTrade(inputCurrencyCode, outputCurrencyCode, inputAmount, inputFilledAmountUsd, outputAmount, outputFilledAmountUsd, CurrencyExchange.mStable);
+    }
+
+    /**
+     * @dev Claims mStable MTA rewards (if `all` is set, unlocks and claims locked rewards).
+     * @param all If locked rewards should be unlocked and claimed.
+     * @param first Index of the first array element to claim. Only applicable if `all` is true. Feed in the second value returned by the savings vault's `unclaimedRewards(address _account)` function.
+     * @param last Index of the last array element to claim. Only applicable if `all` is true. Feed in the third value returned by the savings vault's `unclaimedRewards(address _account)` function.
+     */
+    function claimMStableRewards(bool all, uint256 first, uint256 last) external fundEnabled onlyRebalancer {
+        MStablePoolController.claimRewards(all, first, last);
     }
 }